documents_to_markdown_converter/requirements.txt
mistralai
tqdm
Pillow
python-dotenv